pyproject.toml
[project]
name = "stat-ood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "hydra-core>=1.3.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "wandb>=0.23.1",
]
